TryReadLock usage fails to protect
Because the lock goes out of scope at the end of TryReadLock
the mutex is unlocked and the critical section intended to
be protected is no longer protected. The lock is acquired
and then immediately dropped, allowing writers to come in
and modify data when it is presumed to be protected.

This updates the usage to keep the acquired lock in scope
by returning it. The caller then has it in the current
stack frame and the section is exclusive until the lock
goes out of scope and is released upon destruction.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5431986

Signed-off-by: Vernon Mauery <vmauery@nvidia.com>